TaskViewTime in TaskView integriert
