--- src/main/java/ch/bfh/bti7081/s2017/grey/TodoTimeView.java
@@ -1,5 +1,0 @@
-package ch.bfh.bti7081.s2017.grey;
-
-public class TodoTimeView {
-
-}
